backend/requirements.txt
fastapi==0.87.0
jsons==1.6.3
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
uvicorn[standard]==0.13.4
